Wait for current line text, rather than caret move
